pyproject.toml
[project]
name = "sql-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "faker>=36.1.1",
    "langchain>=0.3.19",
    "langchain-anthropic>=0.3.8",
    "langchain-community>=0.3.18",
    "langchain-google-genai>=2.0.11",
    "langgraph>=0.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
]

[project.scripts]
sql-chain = "sql_chain.graph:main"
init-sql = "sql_chain.sql.initialise:init_database"
test-db = "sql_chain.scripts.test_db:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as an async test",
]
